configure the flashing red light to receive notifications.
In shinken, the redlight gets configured as a notificationway that sends
a simple HTTP request to to the Raspberry pi. When a DOWN or CRITICAL
notification is received, it is saved in a dict with the timestamp.
After the state remains for 30 minutes, or another state notification
(UP, WARNING, OKAY, UKNOWN) is received, the light is turned off.

There is also a web page included to facilitate manual activation of the
red light in times of emergency!